% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pg-get.R
\name{pg.get}
\alias{pg.GET}
\alias{pg.get}
\alias{pgGET}
\alias{pgget}
\alias{postgrest.get}
\title{PostgREST GET}
\usage{
pg.get(domain, table, select = "", filter = "", limit = "",
  pg.filter.syntax = FALSE, encoding = "UTF-8", ...)
}
\arguments{
\item{domain}{The root of the PostgREST server. E.g.
https://postgrest.herokuapp.com}

\item{table}{The table you are querying}

\item{select}{Character vector or comma separated string of columns to return
(optional)}

\item{filter}{Character vector or comma separated string of filter
expressions in R syntax or PostgREST syntax when pg.filter.syntax == TRUE.}

\item{limit}{Integer limiting the number of records returned from the API.}
}
\value{
data.frame of your response
}
\description{
RESTful GET for PostgREST resources.
}
\examples{
pg.get("https://postgrest.herokuapp.com", "speakers", limit = 5, filter = c("id >= 228", "featured == TRUE")))

}

